adding contrib and decisionboundaries visualizer to it
