+++ docs/api/contrib/index.rst
@@ -1,0 +1,23 @@
+.. -*- mode: rst -*-
+
+Contrib
+==========================
+
+The yellowbrick contrib contains a variety of extra tools and experimental visualizers that are outside of core support or are still in development.
+
+
+Classifiers
+==========================
+
+Experimental and undergoing on-going development classification visualizers
+
+.. code:: python
+
+    # Contrib Imports
+
+    from yellowbrick.contrib.classifier import DecisionBoundariesVisualizer
+
+.. toctree::
+   :maxdepth: 2
+
+   boundaries
