+++ tests/baseline_images/test_contrib/test_classifier/test_boundaries/test_integrated_plot_numpy_named_arrays.png
