+++ yellowbrick/classifier/__init__.py
@@ -28,5 +28,8 @@ from .classification_report import ClassificationReport, classification_report
 from .confusion_matrix import ConfusionMatrix
 from .learning_curve import LearningCurveVisualizer, learning_curve_plot
 from .rocauc import ROCAUC, roc_auc
-from .boundaries import decisionviz, DecisionBoundariesVisualizer, DecisionViz
 from .threshold import DiscriminationThreshold, discrimination_threshold
+
+
+# Deprecated and moved to contrib
+from .boundaries import decisionviz, DecisionBoundariesVisualizer, DecisionViz
