+++ yellowbrick/contrib/classifier/__init__.py
@@ -1,0 +1,12 @@
+# yellowbrick.contrib.classifier
+# Visualizations related to evaluating Scikit-Learn classification models
+#
+# Author:   Nathan Danielsen <nathan.danielsen@gmail.com>
+# Created:  Wed Mar 29 12:39:40 2018 -0400
+#
+# Copyright (C) 2016 District Data Labs
+# For license information, see LICENSE.txt
+#
+# ID: __init__.py [5eee25b] nathan.danielsen@gmail.com $
+
+from .boundaries import decisionviz, DecisionBoundariesVisualizer, DecisionViz
